# api/openapi/gateway.yaml
openapi: 3.1.0
info:
  title: AI Gateway API
  version: 0.1.0
  description: OpenAI 风格的数据面、动态管理面，以及 YAML 配置 Schema。
servers:
  - url: http://localhost:8080
    description: 数据面
  - url: http://localhost:9090
    description: 管理面
paths:
  /v1/chat/completions:
    post:
      operationId: createChatCompletion
      security: [{gatewayToken: []}]
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/ChatRequest'}
      responses:
        '200': {$ref: '#/components/responses/ProviderResponse'}
        default: {$ref: '#/components/responses/Error'}
  /v1/responses:
    post:
      operationId: createResponse
      security: [{gatewayToken: []}]
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/ResponsesRequest'}
      responses:
        '200': {$ref: '#/components/responses/ProviderResponse'}
        default: {$ref: '#/components/responses/Error'}
  /v1/images/generations:
    post:
      operationId: generateImage
      security: [{gatewayToken: []}]
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/ImageRequest'}
      responses:
        '200':
          description: 统一 Base64 图片响应
          content:
            application/json:
              schema: {$ref: '#/components/schemas/ImageResponse'}
        default: {$ref: '#/components/responses/Error'}
  /v1/models:
    get:
      operationId: listModels
      security: [{gatewayToken: []}]
      responses:
        '200': {$ref: '#/components/responses/ProviderResponse'}
        default: {$ref: '#/components/responses/Error'}
  /v1/token:
    get:
      operationId: getCurrentToken
      security: [{gatewayToken: []}]
      responses:
        '200':
          description: 当前 Gateway Token 元数据
          content:
            application/json:
              schema: {$ref: '#/components/schemas/GatewayToken'}
        default: {$ref: '#/components/responses/Error'}
...
components:
  securitySchemes:
    gatewayToken:
      type: http
      scheme: bearer
      bearerFormat: agw_opaque
...
  responses:
    ProviderResponse:
      description: 成功响应；具体字段与所调用的 OpenAI 风格接口一致。
      content:
        application/json:
          schema: {type: object, additionalProperties: true}
        text/event-stream:
          schema: {type: string}
    Error:
      description: OpenAI 风格错误
      content:
        application/json:
          schema: {$ref: '#/components/schemas/Error'}
  schemas:
...
    ChatRequest:
      type: object
      required: [model, messages]
      properties:
        model: {type: string, minLength: 1}
        messages: {type: array, minItems: 1, items: {type: object, additionalProperties: true}}
        stream: {type: boolean, default: false}
      additionalProperties: true
    ResponsesRequest:
      type: object
      required: [model, input]
      properties:
        model: {type: string, minLength: 1}
        input: {}
        instructions: {type: string}
        previous_response_id: {type: string}
        max_output_tokens: {type: integer, minimum: 1}
        stream: {type: boolean, default: false}
        background: {const: false}
      additionalProperties: true
    ImageRequest:
      type: object
      required: [model, prompt]
      properties:
        model: {type: string, minLength: 1}
        prompt: {type: string, minLength: 1, maxLength: 32000}
        n: {type: integer, minimum: 1, maximum: 4}
        size: {type: string}
        quality: {type: string}
      additionalProperties: true
    ImageResponse:
      type: object
      required: [created, data]
      properties:
        created: {type: integer, format: int64}
        data:
          type: array
          minItems: 1
          items:
            type: object
            required: [b64_json]
            properties:
              b64_json: {type: string, contentEncoding: base64}
              revised_prompt: {type: string}
            additionalProperties: false
...
    GatewayToken:
      type: object
      required: [id, name, status, created_at, updated_at]
      properties:
        id: {type: string}
        name: {type: string}
        status: {type: string, enum: [active, revoked]}
        created_at: {type: integer, format: int64}
        updated_at: {type: integer, format: int64}
        expires_at: {type: [integer, 'null'], format: int64}
        last_used_at: {type: [integer, 'null'], format: int64}
      additionalProperties: false
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [message, type, code]
          properties:
            message: {type: string}
            type: {const: gateway_error}
            param: {type: ['null', string]}
            code: {type: string}
          additionalProperties: false
      additionalProperties: false